Template Reference Data

A dataset containing a template of input data for the rfs_get_species_list function.

Usage

data_template_ref

Format

A data frame with the following columns:

scientific_name_complete

Complete scientific name of the species.

scientific_name

Scientific name without authorship.

scientific_name_author

Authorship of the scientific name including year.

scientific_name_genus

Genus of the scientific name.

scientific_name_family

Family of the scientific name.

scientific_name_status

Current status of the scientific name (e.g., "Valid", "Synonym", "Uncertain").

species_valid_name_complete

Complete valid name of the species.

species_valid_as_complete

Valid name of the species if it is a synonym.

species_valid_as_short

Short valid name of the species (genus + epithet).

species_valid_name_genus

Genus of the valid name.

species_valid_name_author

Authorship of the valid name including year.

species_valid_name_year

Year of the valid species name.

scientific_name_synonym_of

Valid species of the synonym (only genus + epithet).

all_valid_species_complete

All valid species names associated with the scientific name. Separated by semicolons.

scientific_name_habitat

Habitat information for the scientific name (e. g., "freshwater", "marine").

Source

Organized according to taxonomic data on Cichla.