Download occurrence data for a specific area from GBIF

This function downloads occurrence data for a specific area from GBIF. Obs.: This function is unstable and may not work properly due to time constraints and server limitations while downloading the data. If the first execution did not work properly, please try to rerun the function using the same code. Alternatively, you can use the link, showed during function execution, to download the data manually.

Usage

rfs_download_occ_data(
  gbif_user,
  gbif_pwd,
  gbif_email,
  country = "Brazil",
  state = "Paraná",
  taxonkey = 44,
  user_polygon = NULL,
  max_attempts = 10,
  sleep_time = 120,
  reverse_polygon = FALSE,
  folder_name = "occ_data",
  link = NULL
)

Arguments

gbif_user

GBIF username.

gbif_pwd

GBIF password.

gbif_email

GBIF email.

country

Country name. Default is Brazil.

state

State name. Default is Paraná.

taxonkey

GBIF taxonKey. Default is Chordata (44).

user_polygon

User-defined polygon. Default is NULL.

max_attempts

Maximum number of download attempts. Default is 10

sleep_time

Sleep time between download attempts. Default is 120 s.

reverse_polygon

Logical. If TRUE, the polygon will be reversed.

folder_name

Folder name where the data will be saved. Default is "occ_data".

link

Download link for the occurrence data. Default is NULL.

Value

Downloaded occurrence data.

Examples

if (FALSE) { # \dontrun{
rfs_download_occ_data(gbif_user = "gbif_user",
  gbif_pwd = "gbif_pwd",
  gbif_email = "gbif_email",
  country = "Brazil",
  state = "São Paulo")
} # }