Get the species list from a dataframe based on rFishStatus template

This function uses a dataframe of scientific names from organized a list of species names with their status.

Usage

rfs_get_species_list(df, summary = FALSE)

Arguments

df

A data frame based on the rFishStatus template (rFishStatus::data_template_ref).

Value

A dataframe with the species list.

Examples

df <- rFishStatus::data_template_ref
df_list <- rfs_get_species_list(df)
#> ℹ Creating species list from the input dataset.
#> Looking for uncertain species.
#> Looking for valid species.
#> Looking for all previous valid species names.
#> Filtering valid species.
#> Checking for inconsistencies.
#> Creating final dataset.
#> Adding habitat information.
#> ✔ Species list created successfully.