Check input list against a reference list and perform fuzzy matching

This function takes an input list and a reference list, and performs a series of checks and fuzzy matching operations to create a verified list. The function checks if the input list is a data frame, if it has a column named "species", if the reference list is a data frame, and if the reference list has the mandatory columns. It also allows for specifying the method and maximum distance for fuzzy matching. The function returns a final data frame with the verified list.

Usage

rfs_update_list(
  input_list,
  reference_list,
  complete_name = FALSE,
  method = "lv",
  max_dist = 3,
  max_dist_2 = 5
)

Arguments

input_list

A data frame representing the input list to be checked. Must contain a column named "species". Species name like genus + "sp." should be avoided. If you want to use complete species names (genus + species

authors + year) as key, you should avoid using comma to separate authors and year, and use space instead. For example: 'Genus species Author 0000'.

reference_list

A data frame representing the reference list to compare against. This is obtained using rFishStatus::rfs_get_species_list().

complete_name

A logical value indicating whether to use the species + author + year nomenclature (please, avoid using comma before the year), TRUE, or genus + epithet, FALSE, for joining. Default is FALSE.

method

The method to use for fuzzy matching. Default is "lv" (Levenshtein distance).

max_dist

The maximum distance for fuzzy matching. Default is 3. The user can change this value according to the desired string distance; this will affect the number of matches and false positives, and will vary according to the input list. Most of cases between 2 and 3 are good values when using species names, and 4 to 5 when using complete names. Please, check the fuzzy matching results carefully.

max_dist_2

The maximum dostance for deeper fuzzy matching. You can keep the same as of "max_dist" or increase it by 1 or 2. This can improve query_species solving but also increase false positives.

Value

A data frame containing the final verified list.

Examples

input_list <- data.frame(
   species = c("Acanthochromis polyacanthus", "Acanthochromis sp.", "Cichla kelberi", "Cichla piquitii",
               "Cichla monocolos", "Cichla monoculus", "Cicla monoculus")
)
reference_list <- rFishStatus::rfs_get_species_list(
   rFishStatus::data_template_ref
)
#> ℹ Creating species list from the input dataset.
#> Looking for uncertain species.
#> Looking for valid species.
#> Looking for all previous valid species names.
#> Filtering valid species.
#> Checking for inconsistencies.
#> Creating final dataset.
#> Adding habitat information.
#> ✔ Species list created successfully.
checked_list <- rfs_update_list(input_list, reference_list)
#> ℹ Joining data (exact matching)
#> ℹ Fuzzy matching
#> ℹ Creating verified list